Changed bean-package to model.bean
